Move migration code into feature module dir
- changed some named exports to default
- updated import paths
